Delete unused code
